requirements.txt
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.2
colorama==0.4.6
filelock==3.18.0
Flask==3.1.1
fsspec==2025.7.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
pillow==11.3.0
PyMuPDF==1.26.3
requests==2.32.4
setuptools==80.9.0
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
